tests/integration/slo-test-provider/requirements.txt
ops>=2.0
pydantic>=2,<3
charmlibs-interfaces-sloth @ git+https://github.com/canonical/charmlibs.git@feat/slo#subdirectory=interfaces/sloth
